Fix for OCC infinite loop in ShapeHealing Time out added to ShapeHealing and Boolean Performs

+++ Xbim.Geometry.Engine/XbimGeometryCreator.h
@@ -47,7 +47,7 @@ namespace Xbim
 				Assembly::Load("Xbim.Tessellator");*/
 				
 				String^ timeOut = ConfigurationManager::AppSettings["BooleanTimeOut"];
-				if (!double::TryParse(timeOut, BooleanTimeOut))
+				if (!int::TryParse(timeOut, BooleanTimeOut))
 					BooleanTimeOut = 60;
 				String^ ignoreIfcSweptDiskSolidParamsString = ConfigurationManager::AppSettings["IgnoreIfcSweptDiskSolidParams"];
 				if(!bool::TryParse(ignoreIfcSweptDiskSolidParamsString,IgnoreIfcSweptDiskSolidParams))
@@ -68,7 +68,7 @@ namespace Xbim
 			static void LogError(ILogger^ logger, Object^ entity, String^ format, ... array<Object^>^ arg);
 			static void LogDebug(ILogger^ logger, Object^ entity, String^ format, ... array<Object^>^ arg);
 
-			static double BooleanTimeOut;
+			static int BooleanTimeOut;
 			
 			static bool IgnoreIfcSweptDiskSolidParams;
 			virtual XbimShapeGeometry^ CreateShapeGeometry(IXbimGeometryObject^ geometryObject, double precision, double deflection, double angle, XbimGeometryType storageType, ILogger^ logger);
